# specs/001-enterprise-chat-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: InternalChat API
  version: 1.0.0
  description: >
    HTTP contract for the enterprise internal chat platform.
    All endpoints require a Keycloak-issued bearer token and a resource-scoped
    authorization policy. Errors use Problem Details (RFC 9457).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /conversations/{conversationId}/messages:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Messages]
      summary: Keyset-paginated history (FR-013). Never returns below the caller's visibleFromSeq.
      parameters:
        - name: beforeSeq
          in: query
          schema: { type: integer, format: int64 }
          description: Return messages with seq strictly below this value, newest first.
        - name: afterSeq
          in: query
          schema: { type: integer, format: int64 }
          description: Catch-up after reconnect. Returns messages above this seq, oldest first.
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessagePage" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Messages]
      summary: >
        Send a message (FR-009, FR-011). Idempotent on clientMessageKey — a retry returns
        200 with the existing message rather than creating a duplicate.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "200":
          description: Already accepted under this clientMessageKey (idempotent replay)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: messageId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Messages]
      summary: Edit own message within 24 hours (FR-014)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 8000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "403": { description: Not the author }
        "422": { description: Edit window has expired }
    delete:
      tags: [Messages]
      summary: Delete own message within 24 hours (FR-014). Leaves a visible tombstone.
      responses:
        "204": { description: Deleted }
        "403": { description: Not the author }
        "422": { description: Delete window has expired }
  /conversations/{conversationId}/read-state:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Notifications]
      summary: >
        Advance read position (FR-036). Monotonic — a lower value is ignored, not applied,
        so out-of-order device reports cannot move unread state backwards.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lastReadSeq]
              properties:
                lastReadSeq: { type: integer, format: int64 }
      responses:
        "200":
          description: Effective read state after the monotonic merge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReadState" }
  /conversations/{conversationId}/mute:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Notifications]
      summary: >
        Mute or unmute this conversation for the caller (FR-037). Access is unaffected —
        unread state still updates; only the interrupting notification is suppressed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mutedUntil: { type: string, format: date-time, nullable: true }
      responses:
        "200":
          description: The mute setting now in effect
          content:
            application/json:
              schema:
                type: object
                properties:
                  mutedUntil: { type: string, format: date-time, nullable: true }
...
  /notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get do-not-disturb and digest preferences (FR-037, FR-038)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
    put:
      tags: [Notifications]
      summary: Update preferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationPreferences" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
  /notifications/subscriptions:
    post:
      tags: [Notifications]
      summary: Register this browser for push (FR-034)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushSubscriptionRequest" }
      responses:
        "201": { description: Registered }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: >
        Not a member of this resource. Indistinguishable from 404 in body and timing, so that
        a refusal never reveals that the resource exists (SC-017).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457. Never contains stack traces, SQL, or internal hostnames.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
...
    Message:
      type: object
      required: [id, conversationId, seq, authorId, sentAt]
      properties:
        id: { type: string, format: uuid }
        conversationId: { type: string, format: uuid }
        seq:
          type: integer
          format: int64
          description: Server-assigned, gapless per conversation. The ordering authority.
        authorId: { type: string, format: uuid }
        clientMessageKey: { type: string }
        body: { type: string, nullable: true, description: Null when deleted }
        sentAt: { type: string, format: date-time }
        editedAt: { type: string, format: date-time, nullable: true }
        deletedAt: { type: string, format: date-time, nullable: true }
        mentions:
          type: array
          items: { type: string, format: uuid }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    MessagePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Message" }
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
    SendMessageRequest:
      type: object
      required: [clientMessageKey, body]
      properties:
        clientMessageKey:
          type: string
          description: >
            Client-generated ULID. The uniqueness of (conversationId, clientMessageKey) is what
            makes a retried send idempotent rather than duplicated (FR-011).
        body: { type: string, minLength: 1, maxLength: 8000 }
        attachmentIds:
          type: array
          items: { type: string, format: uuid }
        mentions:
          type: array
          items: { type: string, format: uuid }
    ReadState:
      type: object
      properties:
        conversationId: { type: string, format: uuid }
        lastReadSeq: { type: integer, format: int64 }
        unreadCount: { type: integer }
...
    Attachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [image, video] }
        contentType: { type: string }
        byteSize: { type: integer, format: int64 }
        durationSeconds: { type: integer, nullable: true }
        fileName: { type: string }
        scanStatus: { type: string, enum: [pending, clean, infected, failed] }
        contentUrl:
          type: string
          nullable: true
          description: Present only when scanStatus is clean. Always authorized per request.
        posterUrl: { type: string, nullable: true }
...
    NotificationPreferences:
      type: object
      properties:
        dndStart: { type: string, nullable: true, example: "18:00" }
        dndEnd: { type: string, nullable: true, example: "08:00" }
        timeZone: { type: string, example: "Asia/Ho_Chi_Minh" }
        digestAfterMinutes: { type: integer, default: 60 }
    PushSubscriptionRequest:
      type: object
      required: [endpoint, p256dh, auth]
      properties:
        endpoint: { type: string }
        p256dh: { type: string }
        auth: { type: string }
...